updated added Elastic net regression and reformat
I reformatted and cleaned up the notebook to match the paper better. I can't run seaborn for some reason on my miniconda. Also, let's pare the code under LRM model to exactly what we need, then mirror it with ENR. I tried, and it doesn't work (my error, but I think it is a start).  Once we do that- we will copy the code chunks and run against KNN and RF. I didn't want to create those chunks until we had section 1 finished.